caching for expensive properties
Fetching of customers, projects and services is now cached using
`@functools.cache` decorator. This will save bandwidth, even though it
might reduce some updates. Still, for now it is not expected that
users of this library will use it in long-running scripts, so having
some slightly stale data is not a problem.

Of course, caches can be busted if one so desires. See functools
documentation in the Python standard library for more info.